Move simple message dupe check to --no_unit_tests. Add more elaborate check during runtime.
