+++ Unittest.py
@@ -18,7 +18,7 @@ from Item import ItemInfo
 from ItemPool import remove_junk_items, remove_junk_ludicrous_items, ludicrous_items_base, ludicrous_items_extended, trade_items, ludicrous_exclusions
 from LocationList import location_is_viewable
 from Main import main, resolve_settings, build_world_graphs
-from Messages import Message, check_message_duplicates
+from Messages import Message
 from Settings import Settings, get_preset_files
 
 test_dir = os.path.join(os.path.dirname(__file__), 'tests')
@@ -627,10 +627,6 @@ class TestHints(unittest.TestCase):
         _, spoiler = generate_with_plandomizer("plando-blue-fire-arrows-hints")
         woth = list(spoiler[':woth_locations'].values())
         self.assertIn('Blue Fire Arrows', woth)
-
-class TestMessages(unittest.TestCase):
-    def test_messages_for_duplicates(self):
-        check_message_duplicates()
 
 class TestEntranceRandomizer(unittest.TestCase):
     def test_spawn_point_invalid_areas(self):
